added new examples, fixed transmission, and fixed other bugs

+++ packages/three_js_core/lib/objects/skinned_mesh.dart
@@ -223,7 +223,7 @@ class SkinnedMesh extends Mesh {
       else {
         vector.setValues(1, 0, 0, 0); // do something reasonable
       }
-      skinWeight.setXYZW(i, vector.x.toDouble(), vector.y.toDouble(), vector.z.toDouble(), vector.w.toDouble());
+      skinWeight.setXYZW(i, vector.x, vector.y, vector.z, vector.w);
     }
   }
